+++ CHANGELOG.md
@@ -11,6 +11,13 @@ Prefix your message with one of the following:
 - [Security] in case of vulnerabilities.
 -->
 
+## Unreleased
+
+- [Added] `<meta name="viewport" content="width=device-width,initial-scale=1">`
+  is now added by default.
+- [Changed] You can now group all translations under
+  `page_meta.controller.action.{title,description,keywords}`.
+
 ## 1.0.0
 
 - Initial release.